requirements.txt
Flask==3.0.3
flask-cors==4.0.1
psycopg[binary]==3.2.13
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
